Fixed the CDCL algorithm, implemented EVSIDS and changed how the two watched literals are accessed
The CDCL algorithm didn't use the first UIP for clause learning. This should now be fixed.
The two first literals of a clause are now always the watched literals. If one of the watched
literals turn false, it swaps positions with another none false literal. I also
implemented the EVSIDS variable selection strategy.

+++ NovelSatSolver/src/main/java/cas/thomas/Parsing/ClauseParser.java
@@ -48,7 +48,7 @@ public class ClauseParser {
 
         numberOfVariables++;
         Constraint[] constraints = new Constraint[numberOfClauses];
-        int[] variableOccurences = new int[numberOfVariables];
+        double[] variableOccurences = new double[numberOfVariables];
         SolutionCheckerConstraint[] solutionCheckerConstraints = new SolutionCheckerConstraint[numberOfClauses];
         List<Constraint>[] positivelyWatchedDisjunctiveConstraints = (ArrayList<Constraint>[]) new ArrayList[numberOfVariables];
         List<Constraint>[] negativelyWatchedDisjunctiveConstraints =
@@ -186,7 +186,7 @@ public class ClauseParser {
 
     }
 
-    private void addVariableOccurenceCount(int[] literals, int[] variableOccurences) {
+    private void addVariableOccurenceCount(int[] literals, double[] variableOccurences) {
         for (int i = 0; i < literals.length; i++) {
             variableOccurences[Math.abs(literals[i])] += 1;
         }
